Blended together Cachex.Macros and Cachex.Macros.Boilerplate
